Fixing state label.

+++ Turn Touch Remote/Models/TTDevice.h
@@ -27,7 +27,7 @@ typedef enum {
 @property (nonatomic, readwrite) BOOL isNotified;
 @property (nonatomic, readwrite) BOOL needsReconnection;
 @property (nonatomic, readwrite) BOOL inDFU;
-@property (nonatomic) int firmwareVersion;
+@property (nonatomic) NSInteger firmwareVersion;
 @property (nonatomic) BOOL isFirmwareOld;
 @property (nonatomic) TTDeviceState state;
